# packages/contracts/openapi/laundry-v2.openapi.json
{
  "components": {
    "schemas": {
      "auth.access_session.response": {
        "additionalProperties": false,
        "properties": {
          "access_token": {
            "pattern": "^[A-Za-z0-9_-]+(?:\\.[A-Za-z0-9_-]+){2}$",
            "type": "string"
          },
          "display": {
            "additionalProperties": false,
            "properties": {
              "org_code": {
                "type": "string"
              },
              "staff_name": {
                "type": "string"
              },
              "store_code": {
                "type": "string"
              },
              "store_name": {
                "type": "string"
              }
            },
            "required": [
              "store_name",
              "staff_name",
              "org_code",
              "store_code"
            ],
            "type": "object"
          },
          "expires_in": {
            "const": 900,
            "type": "number"
          },
          "features": {
            "additionalProperties": {
              "type": "boolean"
            },
            "type": "object"
          },
          "role": {
            "enum": [
              "admin",
              "staff"
            ],
            "type": "string"
          },
          "session": {
            "additionalProperties": false,
            "properties": {
              "device_id": {
                "format": "uuid",
                "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
                "type": "string"
              },
              "org_id": {
                "format": "uuid",
                "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
                "type": "string"
              },
              "permission_version": {
                "exclusiveMinimum": 0,
                "maximum": 9007199254740991,
                "type": "integer"
              },
              "session_id": {
                "format": "uuid",
                "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
                "type": "string"
              },
              "session_version": {
                "exclusiveMinimum": 0,
                "maximum": 9007199254740991,
                "type": "integer"
              },
              "staff_id": {
                "format": "uuid",
                "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
                "type": "string"
              },
              "store_id": {
                "format": "uuid",
                "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
                "type": "string"
              }
            },
            "required": [
              "session_id",
              "session_version",
              "org_id",
              "store_id",
              "staff_id",
              "device_id",
              "permission_version"
            ],
            "type": "object"
          },
          "storage": {
            "const": "memory_only",
            "type": "string"
          },
          "token_type": {
            "const": "Bearer",
            "type": "string"
          }
        },
        "required": [
          "access_token",
          "token_type",
          "expires_in",
          "storage",
          "session",
          "role",
          "features",
          "display"
        ],
        "type": "object"
      },
      "auth.empty.request": {
        "additionalProperties": false,
        "properties": {},
        "type": "object"
      },
      "auth.login.request": {
        "additionalProperties": false,
        "properties": {
          "device_id": {
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
            "type": "string"
          },
          "org_code": {
            "pattern": "^[\\x21-\\x7E]{1,128}$",
            "type": "string"
          },
          "password": {
            "maxLength": 1024,
            "minLength": 1,
            "type": "string"
          },
          "store_code": {
            "pattern": "^[\\x21-\\x7E]{1,128}$",
            "type": "string"
          },
          "username": {
            "pattern": "^[\\x21-\\x7E]{1,128}$",
            "type": "string"
          }
        },
        "required": [
          "org_code",
          "store_code",
          "username",
          "password",
          "device_id"
        ],
        "type": "object"
      },
      "auth.logout.response": {
        "additionalProperties": false,
        "properties": {
          "logged_out": {
            "const": true,
            "type": "boolean"
          }
        },
        "required": [
          "logged_out"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "bearerFormat": "JWT",
        "description": "Memory-only access token (never store in cookies or Web Storage).",
        "scheme": "bearer",
        "type": "http"
      },
      "csrfHeader": {
        "description": "Double-submit CSRF header paired with the readable CSRF cookie.",
        "in": "header",
        "name": "x-csrf-token",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "Contract-first OpenAPI 3.1 projection for frozen M1 and M2 counter contracts. Auth paths come solely from AUTH_OPERATION_MATRIX. Bus commands/queries come from M1_FIRST_WAVE_DEFINITIONS and M2_CONTRACT_DEFINITIONS. Errors use the A2 CommandResponse / CommandFailureResponse envelope.",
    "title": "Laundry Desk v2 API",
    "version": "0.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v2/auth/login": {
      "post": {
        "description": "Browser auth operation `login` (ingress: lifecycle_http). Registry command: `identity.login`. Allowed public errors: VALIDATION_FAILED, AUTHENTICATION_FAILED, CSRF_REJECTED, RATE_LIMITED, TRANSACTION_FAILED, EVENT_DISPATCH_FAILED. Projection source: AUTH_OPERATION_MATRIX (A5/A7). Server-only hashes and factories are omitted.",
        "operationId": "auth_login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/auth.login.request"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/auth.access_session.response"
                }
              }
            },
            "description": "Successful browser auth response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandFailureResponse"
                }
              }
            },
            "description": "Authentication failed"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandFailureResponse"
                }
              }
            },
            "description": "Request origin verification failed"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandFailureResponse"
                }
              }
            },
            "description": "Too many requests"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandFailureResponse"
                }
              }
            },
            "description": "Unified A2 command failure envelope"
          }
        },
        "summary": "Auth login",
        "tags": [
          "auth"
        ],
        "x-laundry-kind": "auth"
      }
    },
    "/api/v2/auth/logout": {
      "post": {
        "description": "Browser auth operation `logout` (ingress: lifecycle_http). Registry command: `identity.logout`. Allowed public errors: VALIDATION_FAILED, AUTHENTICATION_FAILED, CSRF_REJECTED, RATE_LIMITED, TRANSACTION_FAILED, EVENT_DISPATCH_FAILED. Projection source: AUTH_OPERATION_MATRIX (A5/A7). Server-only hashes and factories are omitted.",
        "operationId": "auth_logout",
        "parameters": [
          {
            "description": "Double-submit CSRF proof; must match the readable CSRF cookie value.",
            "in": "header",
            "name": "x-csrf-token",
            "required": true,
            "schema": {
              "pattern": "^v1\\.[A-Za-z0-9_-]{43,128}$",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/auth.empty.request"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/auth.logout.response"
                }
              }
            },
            "description": "Successful browser auth response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandFailureResponse"
                }
              }
            },
            "description": "Authentication failed"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandFailureResponse"
                }
              }
            },
            "description": "Request origin verification failed"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandFailureResponse"
                }
              }
            },
            "description": "Too many requests"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandFailureResponse"
                }
              }
            },
            "description": "Unified A2 command failure envelope"
          }
        },
        "security": [
          {
            "csrfHeader": []
          }
        ],
        "summary": "Auth logout",
        "tags": [
          "auth"
        ],
        "x-laundry-kind": "auth"
      }
    },
...
    "/api/v2/auth/refresh": {
      "post": {
        "description": "Browser auth operation `refresh` (ingress: lifecycle_http). Registry command: `identity.refresh`. Allowed public errors: VALIDATION_FAILED, AUTHENTICATION_FAILED, CSRF_REJECTED, RATE_LIMITED, TRANSACTION_FAILED, EVENT_DISPATCH_FAILED. Projection source: AUTH_OPERATION_MATRIX (A5/A7). Server-only hashes and factories are omitted.",
        "operationId": "auth_refresh",
        "parameters": [
          {
            "description": "Double-submit CSRF proof; must match the readable CSRF cookie value.",
            "in": "header",
            "name": "x-csrf-token",
            "required": true,
            "schema": {
              "pattern": "^v1\\.[A-Za-z0-9_-]{43,128}$",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/auth.empty.request"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/auth.access_session.response"
                }
              }
            },
            "description": "Successful browser auth response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandFailureResponse"
                }
              }
            },
            "description": "Authentication failed"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandFailureResponse"
                }
              }
            },
            "description": "Request origin verification failed"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandFailureResponse"
                }
              }
            },
            "description": "Too many requests"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandFailureResponse"
                }
              }
            },
            "description": "Unified A2 command failure envelope"
          }
        },
        "security": [
          {
            "csrfHeader": []
          }
        ],
        "summary": "Auth refresh",
        "tags": [
          "auth"
        ],
        "x-laundry-kind": "auth"
      }
    },
...
  }
}